Selecting the optimal robustification parameter tau for the element-wise truncated or adaptive Huber estimators

As the title

optim.tuning(
  X,
  S,
  h = 1,
  length_tau = 5,
  M_est = FALSE,
  max.tau = 10,
  start.prob = 0.97,
  ...
)

Arguments

X
A numeric pxn matrix containing n observations of a p-dimensional time series.

S
A strictly positive integer corresponding to the chosen number of non-overlapping blocks.

h
A strictly positive integer corresponding to how many neighboring blocks will be removed.

length_tau
A strictly positive integer corresponding to the length of the vector of candidate robustification parameters.

M_est
A logical scalar indicating which truncated estimator will be used (F: truncated mean estimator; T: adaptive Huber M-estimator).

max.tau
A positive numeric scalar corresponding to the maximum value of the candidate robustification parameters.

start.prob
A numeric scalar in (0,1) such that the minimum value of the candidate robustification parameters is the empirical "start.prob" quantile of the corresponding coordiante of X.

...
Additional arguments.

Value

A numeric vector with each entry corresponds the selected robustification parameter for for each coordinate.

Author

Haotian Xu and Stephane Guerrier

Examples

n = 100
p = 50
Sigma = diag(1, p)
set.seed(123)
X = VAR1_simu(n = n, mu = rep(0, p), skip = 300, Sigma.mat = Sigma, rho = 0.7, err.dist = "pareto")
optim.tuning(X, S = 10, h = 1, M_est = FALSE, max.tau = 100)

#>  [1]  2.762182  2.081750  1.328890  3.447785  3.405271  1.912649  3.709577
#>  [8]  6.961192  2.308370  2.483048  1.468468  2.618611  1.730653  3.003648
#> [15]  1.758279  2.675418  4.825309  5.041755  4.651768  1.865297  3.052670
#> [22]  2.164891  2.841551  2.523507  1.446313  4.341307  1.118869  4.830270
#> [29]  8.487456  1.358423  3.956177  1.422615  2.498233  2.380974  2.448530
#> [36]  2.035460  2.131922  1.992328  3.738890  1.502218  5.082644  1.450227
#> [43]  2.089618  2.379843  1.421194 11.945274  4.007887  2.236078  2.026902
#> [50]  1.698556