Block-wise cross-validation to select the spectrum-wise robustification parameter tau for autocovariance

As the title

spec_block_cv(X, S, h, lag, tau.vec, ...)

Arguments

X
A pxn matrix of observated multivariate time series.

S
A strictly positive integer corresponding to the chosen number of non-overlapping blocks.

h
A strictly positive integer corresponding to how many neighboring blocks will be removed.

lag
A strictly non-negative integer corresponding to the lag of the autocovariance.

tau.vec
A numeric vector of candidate robustification parameters (positive values).

...
Additional arguments.

Value

A numeric vector of cross validation errors (in l2 norm), corresponding to the tau.vec.

Author

Haotian Xu and Stephane Guerrier

Examples

n = 100
p = 50
Sigma = diag(1, p)
set.seed(123)
X = VAR1_simu(n = n, mu = rep(0, p), skip = 300, Sigma.mat = Sigma, rho = 0.7, err.dist = "pareto")
tau.vec = seq(from = 10, to = 100, length.out = 10)
spec_block_cv(X, S = 10, h = 1, lag = 1, tau.vec)

#>           [,1]
#>  [1,] 58.67065
#>  [2,] 58.66713
#>  [3,] 58.78328
#>  [4,] 59.01085
#>  [5,] 59.30613
#>  [6,] 59.63504
#>  [7,] 59.94445
#>  [8,] 60.23879
#>  [9,] 60.52151
#> [10,] 60.78827